mcp-server-client/requirements.txt
beautifulsoup4==4.13.4
httpx==0.28.1
mcp==1.8.0
openai==1.78.0
python-dotenv==1.1.0
starlette==0.46.2
uvicorn==0.34.2
